Pickles_Auctions/requirements.txt
astroid==2.6.5
colorama==0.4.4
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
numpy==1.19.5
pandas==1.1.5
pylint==2.9.6
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
flask